+++ shared/constant/constants.go
@@ -23,9 +23,6 @@ type Assistant string         // Always "assistant"
 type BookAppointment string   // Always "book_appointment"
 type CheckAvailability string // Always "check_availability"
 type Comparative string       // Always "comparative"
-type Deepgram string          // Always "Deepgram"
-type DeepgramNova2 string     // Always "deepgram/nova-2"
-type DeepgramNova3 string     // Always "deepgram/nova-3"
 type Developer string         // Always "developer"
 type Function string          // Always "function"
 type Handoff string           // Always "handoff"
@@ -49,9 +46,6 @@ func (c Assistant) Default() Assistant                 { return "assistant" }
 func (c BookAppointment) Default() BookAppointment     { return "book_appointment" }
 func (c CheckAvailability) Default() CheckAvailability { return "check_availability" }
 func (c Comparative) Default() Comparative             { return "comparative" }
-func (c Deepgram) Default() Deepgram                   { return "Deepgram" }
-func (c DeepgramNova2) Default() DeepgramNova2         { return "deepgram/nova-2" }
-func (c DeepgramNova3) Default() DeepgramNova3         { return "deepgram/nova-3" }
 func (c Developer) Default() Developer                 { return "developer" }
 func (c Function) Default() Function                   { return "function" }
 func (c Handoff) Default() Handoff                     { return "handoff" }
@@ -75,9 +69,6 @@ func (c Assistant) MarshalJSON() ([]byte, error)         { return marshalString(
 func (c BookAppointment) MarshalJSON() ([]byte, error)   { return marshalString(c) }
 func (c CheckAvailability) MarshalJSON() ([]byte, error) { return marshalString(c) }
 func (c Comparative) MarshalJSON() ([]byte, error)       { return marshalString(c) }
-func (c Deepgram) MarshalJSON() ([]byte, error)          { return marshalString(c) }
-func (c DeepgramNova2) MarshalJSON() ([]byte, error)     { return marshalString(c) }
-func (c DeepgramNova3) MarshalJSON() ([]byte, error)     { return marshalString(c) }
 func (c Developer) MarshalJSON() ([]byte, error)         { return marshalString(c) }
 func (c Function) MarshalJSON() ([]byte, error)          { return marshalString(c) }
 func (c Handoff) MarshalJSON() ([]byte, error)           { return marshalString(c) }